backend/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sse-starlette==1.6.5

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
mysqlclient==2.2.0
alembic==1.12.1

# Pydantic Settings
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# LangChain & RAG
langchain==0.0.340
langchain-community==0.0.10
chromadb==0.4.22
sentence-transformers==2.2.2
tiktoken==0.5.1

# LLM Integration
openai==1.3.0
groq==0.4.2

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
markdown==3.5.1
beautifulsoup4==4.12.2

# Utilities
numpy==1.24.3
pandas==2.1.3
tqdm==4.66.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Testing & Development
pytest==7.4.3
httpx==0.25.1
black==23.11.0
isort==5.12.0
flake8==6.1.0